Fix a possibly overflow with the CLUT size of DVB and PGS subtitles.
Because of this PGS subtitles were sometimes not displayed at all, the impact was probably negligible for DVB subtitles.

Fixes #5515.

+++ src/Subtitles/DVBSub.h
@@ -75,7 +75,7 @@ private:
     struct DVB_CLUT {
         BYTE id = 0;
         BYTE version_number = 0;
-        BYTE size = 0;
+        WORD size = 0;
 
         std::array<HDMV_PALETTE, 256> palette;
